+++ custom_components/mydolphin_plus/sensor.py
@@ -1,6 +1,6 @@
 import logging
 
-from homeassistant.components.sensor import SensorEntity, SensorEntityDescription
+from homeassistant.components.sensor import SensorEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import ATTR_ICON, ATTR_STATE, Platform
 from homeassistant.core import HomeAssistant, callback
@@ -8,11 +8,10 @@ from homeassistant.helpers.dispatcher import async_dispatcher_connect
 
 from .common.base_entity import MyDolphinPlusBaseEntity, async_setup_entities
 from .common.consts import ATTR_ATTRIBUTES, SIGNAL_DEVICE_NEW
+from .common.entity_descriptions import MyDolphinPlusSensorEntityDescription
 from .managers.coordinator import MyDolphinPlusCoordinator
 
 _LOGGER = logging.getLogger(__name__)
-
-CURRENT_DOMAIN = Platform.SENSOR
 
 
 async def async_setup_entry(
@@ -26,8 +25,7 @@ async def async_setup_entry(
         async_setup_entities(
             hass,
             entry,
-            CURRENT_DOMAIN,
-            SensorEntityDescription,
+            Platform.SENSOR,
             MyDolphinPlusSensorEntity,
             async_add_entities,
         )
@@ -42,10 +40,10 @@ class MyDolphinPlusSensorEntity(MyDolphinPlusBaseEntity, SensorEntity):
 
     def __init__(
         self,
-        entity_description: SensorEntityDescription,
+        entity_description: MyDolphinPlusSensorEntityDescription,
         coordinator: MyDolphinPlusCoordinator,
     ):
-        super().__init__(entity_description, coordinator, CURRENT_DOMAIN)
+        super().__init__(entity_description, coordinator)
 
         self._attr_device_class = entity_description.device_class